Mount CacheRegion ConfigMap in sidecar pods
A CacheRegion ConfigMap is created in user pod namespace, containing all
current CacheRegion configurations associated with the CacheService
defined on the user pod.

A seperate controller sets the OwnerRef of the ConfigMap to the newly
created pod so that the ConfigMap is automatically removed when no
dependent pods exist. It's not possible for this ref to be set in the
webhook itself, as the pod has not yet been created and so does not have
a UUID.

Updates to a CacheRegion spec are propogated to existing ConfigMaps by
the cacheregion_controller. Kubernetes updates the filesystem of existing
user pods so that the latest ConfigMap content is available to the
sidecar container.